backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0

# Parsing
beautifulsoup4==4.12.3
lxml==5.3.0
# AI stack
langchain==1.2.15
langchain_gigachat==0.5.0

# OCR stack
numpy==1.26.4
paddlepaddle==3.3.1
paddleocr==3.4.1
pdf2image==1.17.0
docx==0.2.4
python-docx==1.2.0

# Streamlit demo app
streamlit==1.39.0
pandas==2.2.3
